moved utility fns to separate document
