+++ server/database/data/data-generator.js
@@ -5,6 +5,7 @@
 const faker = require('faker');
 const moment = require('moment');
 const createCsvWriter = require('csv-writer').createObjectCsvWriter;
+const Utils = require('./data-generator-utils');
 
 const numOfHostels = 100000;
 const numOfAuthors = 2000;
@@ -14,43 +15,6 @@ const reviewsFilePath = './data/reviews.csv';
 const authorsFilePath = './data/authors.csv';
 
 const updateThreshold = 10000;
-
-/* Utility Functions */
-const getRandomInt = (min, max) => {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-  return Math.floor(Math.random() * (max - min) + min);
-};
-
-// creates a random decimal to one place value, inclusive of min and max
-const getRandomDec = (min, max) => {
-  max = (max + 1) * 10;
-  min *= 10;
-  return Math.min(Math.floor(Math.random() * (max - min) + min) / 10, 10);
-};
-
-// generates a random date in the last two-ish years
-const getRandomDate = () => {
-  const date = new Date(+(new Date()) - Math.random() * 31556952000);
-  return moment(date).format('YYYY-MM-DD');
-};
-
-// takes in an array of attributes
-// constructs the array of column names that csv-writer needs
-const generateHeader = (attributes) => {
-  const header = [];
-  attributes.forEach((columnName) => {
-    header.push({ id: `${columnName}`, title: `${columnName}` });
-  });
-  return header;
-};
-
-// console logs regular updates
-const giveUpdate = (count, name) => {
-  if (name === 'author') console.log(`📝 ${count} ${name} records generated! 📝`);
-  else if (name === 'hostel') console.log(`🛏 ${count} ${name} records generated! 🛏`);
-  else console.log(`🚀 ${count} ${name} records generated! 🚀`);
-};
 
 /* Data Generation */
 
@@ -76,13 +40,13 @@ async function generateHostels(num) {
       await csvWriter.writeRecords(record)
         .then(() => {
           count += 1;
-          if (count % updateThreshold === 0) giveUpdate(count, 'hostel');
+          if (count % updateThreshold === 0) Utils.giveUpdate(count, 'hostel');
         });
     } catch (error) {
       console.log('an error occurred on record ', i, error);
     }
   }
-  giveUpdate(count, 'hostel');
+  Utils.giveUpdate(count, 'hostel');
 }
 
 // takes in the number of hostels
@@ -92,7 +56,7 @@ async function generateReviews(num) {
   const attributes = ['review_id', 'hostel_id', 'author_id', 'description', 'date', 'security', 'location',
     'staff', 'atmosphere', 'cleanliness', 'facilities', 'value', 'total'];
 
-  const header = generateHeader(attributes);
+  const header = Utils.generateHeader(attributes);
 
   // define the headers
   const csvWriter = createCsvWriter({ path: reviewsFilePath, header });
@@ -101,7 +65,7 @@ async function generateReviews(num) {
   // iterate over hostel ids
   for (let i = 0; i < num; i += 1) {
     // generate a random number of reviews between 1 & 10
-    const numOfReviews = getRandomInt(1, 10);
+    const numOfReviews = Utils.getRandomInt(1, 10);
 
     // generate that many reviews
     for (let j = 0; j < numOfReviews; j += 1) {
@@ -110,16 +74,16 @@ async function generateReviews(num) {
         let record = {
           review_id: count,
           hostel_id: i,
-          author_id: getRandomInt(1, numOfAuthors),
+          author_id: Utils.getRandomInt(1, numOfAuthors),
           description: faker.lorem.paragraph(),
-          date: getRandomDate(),
+          date: Utils.getRandomDate(),
         };
         // loop through the rest of the attributes array
         // add all ratings attributes except total to the record
         let ratingsTotal = 0;
         let ratingsCount = 0;
         for (let k = 5; k < attributes.length - 1; k += 1) {
-          const rating = getRandomDec(1, 10);
+          const rating = Utils.getRandomDec(1, 10);
           record[attributes[k]] = rating;
           // keep track of the running total and count
           ratingsTotal += rating;
@@ -135,14 +99,14 @@ async function generateReviews(num) {
         await csvWriter.writeRecords(record)
           .then(() => {
             count += 1;
-            if (count % updateThreshold === 0) giveUpdate(count, 'review');
+            if (count % updateThreshold === 0) Utils.giveUpdate(count, 'review');
           });
       } catch (error) {
         console.log(`an error occurred on hostel ${i}, review ${j}`, error);
       }
     }
   }
-  giveUpdate(count, 'review');
+  Utils.giveUpdate(count, 'review');
 }
 
 // takes in the number of authors
@@ -154,7 +118,7 @@ async function generateAuthors(num) {
   const userDescriptions = ['Globetrotter', 'Avid Traveller', 'Novice Nomad'];
   const genders = ['Female', 'Male'];
 
-  const header = generateHeader(attributes);
+  const header = Utils.generateHeader(attributes);
 
   // define the headers
   const csvWriter = createCsvWriter({ path: authorsFilePath, header });
@@ -168,21 +132,21 @@ async function generateAuthors(num) {
         {
           author_id: i,
           name: faker.internet.userName(),
-          age_group: userAges[getRandomInt(0, 4)],
-          gender: genders[getRandomInt(0, 2)],
-          authdescription: userDescriptions[getRandomInt(0, 3)],
+          age_group: userAges[Utils.getRandomInt(0, 4)],
+          gender: genders[Utils.getRandomInt(0, 2)],
+          authdescription: userDescriptions[Utils.getRandomInt(0, 3)],
         },
       ];
       await csvWriter.writeRecords(record)
         .then(() => {
           count += 1;
-          if (count % updateThreshold === 0) giveUpdate(count, 'author');
+          if (count % updateThreshold === 0) Utils.giveUpdate(count, 'author');
         });
     } catch (error) {
       console.log('an error occurred on record ', i, error);
     }
   }
-  giveUpdate(count, 'author');
+  Utils.giveUpdate(count, 'author');
 }
 
 generateHostels(numOfHostels);
